+++ core/dbt/adapters/sql/impl.py
@@ -4,7 +4,6 @@ from typing import Any, Optional, Tuple, Type, List
 import dbt.clients.agate_helper
 from dbt.contracts.connection import Connection
 import dbt.exceptions
-import dbt.flags
 from dbt.adapters.base import BaseAdapter, available
 from dbt.adapters.sql import SQLConnectionManager
 from dbt.logger import GLOBAL_LOGGER as logger
