+++ core/dbt/clients/jinja.py
@@ -30,7 +30,7 @@ from dbt.exceptions import (
     InternalException, raise_compiler_error, CompilationException,
     invalid_materialization_argument, MacroReturn
 )
-from dbt.flags import MACRO_DEBUGGING
+from dbt import flags
 from dbt.logger import GLOBAL_LOGGER as logger  # noqa
 
 
@@ -93,8 +93,8 @@ class MacroFuzzEnvironment(jinja2.sandbox.SandboxedEnvironment):
         If the value is 'write', also write the files to disk.
         WARNING: This can write a ton of data if you aren't careful.
         """
-        if filename == '<template>' and MACRO_DEBUGGING:
-            write = MACRO_DEBUGGING == 'write'
+        if filename == '<template>' and flags.MACRO_DEBUGGING:
+            write = flags.MACRO_DEBUGGING == 'write'
             filename = _linecache_inject(source, write)
 
         return super()._compile(source, filename)  # type: ignore
