+++ core/dbt/task/printer.py
@@ -1,58 +1,15 @@
-import textwrap
-import time
 from typing import Dict, Optional, Tuple
-from dbt.logger import GLOBAL_LOGGER as logger, DbtStatusMessage, TextOnly
+from dbt.logger import (
+    GLOBAL_LOGGER as logger,
+    DbtStatusMessage,
+    TextOnly,
+    get_timestamp,
+)
 from dbt.node_types import NodeType
-import dbt.ui.colors
 
-USE_COLORS = False
-
-COLOR_FG_RED = dbt.ui.colors.COLORS['red']
-COLOR_FG_GREEN = dbt.ui.colors.COLORS['green']
-COLOR_FG_YELLOW = dbt.ui.colors.COLORS['yellow']
-COLOR_RESET_ALL = dbt.ui.colors.COLORS['reset_all']
-
-PRINTER_WIDTH = 80
-
-
-def use_colors():
-    global USE_COLORS
-    USE_COLORS = True
-
-
-def printer_width(printer_width):
-    global PRINTER_WIDTH
-    PRINTER_WIDTH = printer_width
-
-
-def get_timestamp():
-    return time.strftime("%H:%M:%S")
-
-
-def color(text: str, color_code: str):
-    if USE_COLORS:
-        return "{}{}{}".format(color_code, text, COLOR_RESET_ALL)
-    else:
-        return text
-
-
-def green(text: str):
-    return color(text, COLOR_FG_GREEN)
-
-
-def yellow(text: str):
-    return color(text, COLOR_FG_YELLOW)
-
-
-def red(text: str):
-    return color(text, COLOR_FG_RED)
-
-
-def print_timestamped_line(msg: str, use_color: Optional[str] = None):
-    if use_color is not None:
-        msg = color(msg, use_color)
-
-    logger.info("{} | {}".format(get_timestamp(), msg))
+from dbt.tracking import InvocationProcessor
+from dbt import ui
+from dbt import utils
 
 
 def print_fancy_output_line(
@@ -68,8 +25,8 @@ def print_fancy_output_line(
         progress=progress,
         message=msg)
 
-    truncate_width = PRINTER_WIDTH - 3
-    justified = prefix.ljust(PRINTER_WIDTH, ".")
+    truncate_width = ui.PRINTER_WIDTH - 3
+    justified = prefix.ljust(ui.PRINTER_WIDTH, ".")
     if truncate and len(justified) > truncate_width:
         justified = justified[:truncate_width] + '...'
 
@@ -99,7 +56,7 @@ def get_counts(flat_nodes) -> str:
         counts[t] = counts.get(t, 0) + 1
 
     stat_line = ", ".join(
-        [dbt.utils.pluralize(v, k) for k, v in counts.items()])
+        [utils.pluralize(v, k) for k, v in counts.items()])
 
     return stat_line
 
@@ -119,7 +76,7 @@ def print_hook_end_line(
 ) -> None:
     msg = 'OK hook: {}'.format(statement)
     # hooks don't fail into this path, so always green
-    print_fancy_output_line(msg, green(status), index, total,
+    print_fancy_output_line(msg, ui.green(status), index, total,
                             execution_time=execution_time, truncate=True)
 
 
@@ -127,21 +84,21 @@ def print_skip_line(
     model, schema: str, relation: str, index: int, num_models: int
 ) -> None:
     msg = 'SKIP relation {}.{}'.format(schema, relation)
-    print_fancy_output_line(msg, yellow('SKIP'), index, num_models)
+    print_fancy_output_line(msg, ui.yellow('SKIP'), index, num_models)
 
 
 def print_cancel_line(model) -> None:
     msg = 'CANCEL query {}'.format(model)
-    print_fancy_output_line(msg, red('CANCEL'), index=None, total=None)
+    print_fancy_output_line(msg, ui.red('CANCEL'), index=None, total=None)
 
 
 def get_printable_result(result, success: str, error: str) -> Tuple[str, str]:
     if result.error is not None:
         info = 'ERROR {}'.format(error)
-        status = red(result.status)
+        status = ui.red(result.status)
     else:
         info = 'OK {}'.format(success)
-        status = green(result.status)
+        status = ui.green(result.status)
 
     return info, status
 
@@ -153,18 +110,18 @@ def print_test_result_line(
 
     if result.error is not None:
         info = "ERROR"
-        color = red
+        color = ui.red
     elif result.status == 0:
         info = 'PASS'
-        color = green
+        color = ui.green
 
     elif result.warn:
         info = 'WARN {}'.format(result.status)
-        color = yellow
+        color = ui.yellow
 
     elif result.fail:
         info = 'FAIL {}'.format(result.status)
-        color = red
+        color = ui.red
 
     else:
         raise RuntimeError("unexpected status: {}".format(result.status))
@@ -227,16 +184,16 @@ def print_seed_result_line(result, schema_name: str, index: int, total: int):
 def print_freshness_result_line(result, index: int, total: int) -> None:
     if result.error:
         info = 'ERROR'
-        color = red
+        color = ui.red
     elif result.status == 'error':
         info = 'ERROR STALE'
-        color = red
+        color = ui.red
     elif result.status == 'warn':
         info = 'WARN'
-        color = yellow
+        color = ui.yellow
     else:
         info = 'PASS'
-        color = green
+        color = ui.green
 
     if hasattr(result, 'node'):
         source_name = result.node.source_name
@@ -298,11 +255,11 @@ def print_run_result_error(
 
     if result.fail or (is_warning and result.warn):
         if is_warning:
-            color = yellow
+            color = ui.yellow
             info = 'Warning'
             logger_fn = logger.warning
         else:
-            color = red
+            color = ui.red
             info = 'Failure'
             logger_fn = logger.error
         logger_fn(color("{} in {} {} ({})").format(
@@ -316,7 +273,7 @@ def print_run_result_error(
         except ValueError:
             logger.error("  Status: {}".format(result.status))
         else:
-            status = dbt.utils.pluralize(result.status, 'result')
+            status = utils.pluralize(result.status, 'result')
             logger.error("  Got {}, expected 0.".format(status))
 
         if result.node.build_path is not None:
@@ -329,7 +286,7 @@ def print_run_result_error(
         first = True
         for line in result.error.split("\n"):
             if first:
-                logger.error(yellow(line))
+                logger.error(ui.yellow(line))
                 first = False
             else:
                 logger.error(line)
@@ -340,24 +297,24 @@ def print_skip_caused_by_error(
 ) -> None:
     msg = ('SKIP relation {}.{} due to ephemeral model error'
            .format(schema, relation))
-    print_fancy_output_line(msg, red('ERROR SKIP'), index, num_models)
+    print_fancy_output_line(msg, ui.red('ERROR SKIP'), index, num_models)
     print_run_result_error(result, newline=False)
 
 
 def print_end_of_run_summary(
     num_errors: int, num_warnings: int, keyboard_interrupt: bool = False
 ) -> None:
-    error_plural = dbt.utils.pluralize(num_errors, 'error')
-    warn_plural = dbt.utils.pluralize(num_warnings, 'warning')
+    error_plural = utils.pluralize(num_errors, 'error')
+    warn_plural = utils.pluralize(num_warnings, 'warning')
     if keyboard_interrupt:
-        message = yellow('Exited because of keyboard interrupt.')
+        message = ui.yellow('Exited because of keyboard interrupt.')
     elif num_errors > 0:
-        message = red("Completed with {} and {}:".format(
+        message = ui.red("Completed with {} and {}:".format(
             error_plural, warn_plural))
     elif num_warnings > 0:
-        message = yellow('Completed with {}:'.format(warn_plural))
+        message = ui.yellow('Completed with {}:'.format(warn_plural))
     else:
-        message = green('Completed successfully')
+        message = ui.green('Completed successfully')
 
     with TextOnly():
         logger.info('')
@@ -365,9 +322,6 @@ def print_end_of_run_summary(
 
 
 def print_run_end_messages(results, keyboard_interrupt: bool = False) -> None:
-    # Prevent import loop from happening by importing tracking here.
-    from dbt.tracking import InvocationProcessor  # noqa
-
     errors = [r for r in results if r.error is not None or r.fail]
     warnings = [r for r in results if r.warn]
     with DbtStatusMessage(), InvocationProcessor():
@@ -382,29 +336,3 @@ def print_run_end_messages(results, keyboard_interrupt: bool = False) -> None:
             print_run_result_error(warning, is_warning=True)
 
         print_run_status_line(results)
-
-
-def line_wrap_message(
-    msg: str, subtract: int = 0, dedent: bool = True, prefix: str = ''
-) -> str:
-    '''
-    Line wrap the given message to PRINTER_WIDTH - {subtract}. Convert double
-    newlines to newlines and avoid calling textwrap.fill() on them (like
-    markdown)
-    '''
-    width = PRINTER_WIDTH - subtract
-    if dedent:
-        msg = textwrap.dedent(msg)
-
-    if prefix:
-        msg = f'{prefix}{msg}'
-
-    # If the input had an explicit double newline, we want to preserve that
-    # (we'll turn it into a single line soon). Support windows, too.
-    splitter = '\r\n\r\n' if '\r\n\r\n' in msg else '\n\n'
-    chunks = msg.split(splitter)
-    return '\n'.join(textwrap.fill(chunk, width=width) for chunk in chunks)
-
-
-def warning_tag(msg: str) -> str:
-    return f'[{yellow("WARNING")}]: {msg}'
