+++ core/dbt/task/snapshot.py
@@ -1,6 +1,18 @@
-from dbt.node_runners import SnapshotRunner
+from .run import ModelRunner, RunTask
+from .printer import print_snapshot_result_line
+
 from dbt.node_types import NodeType
-from dbt.task.run import RunTask
+
+class SnapshotRunner(ModelRunner):
+    def describe_node(self):
+        return "snapshot {}".format(self.get_node_representation())
+
+    def print_result_line(self, result):
+        print_snapshot_result_line(
+            result,
+            self.get_node_representation(),
+            self.node_index,
+            self.num_nodes)
 
 
 class SnapshotTask(RunTask):
