+++ core/dbt/utils.py
@@ -17,10 +17,7 @@ from typing import (
 
 import dbt.exceptions
 
-from dbt.logger import GLOBAL_LOGGER as logger
 from dbt.node_types import NodeType
-from dbt.clients import yaml_helper
-from dbt.ui import printer
 
 DECIMALS: Tuple[Type[Any], ...]
 try:
@@ -65,14 +62,6 @@ def get_model_name_or_none(model):
     else:
         name = str(model)
     return name
-
-
-def compiler_warning(model, msg, resource_type='model'):
-    name = get_model_name_or_none(model)
-    logger.info(
-        "* Compilation warning while compiling {} {}:\n* {}\n"
-        .format(resource_type, name, msg)
-    )
 
 
 MACRO_PREFIX = 'dbt_macro__'
@@ -297,70 +286,6 @@ class memoized:
     def __get__(self, obj, objtype):
         '''Support instance methods.'''
         return functools.partial(self.__call__, obj)
-
-
-def invalid_ref_fail_unless_test(node, target_model_name,
-                                 target_model_package, disabled):
-
-    if node.resource_type == NodeType.Test:
-        msg = dbt.exceptions.get_target_not_found_or_disabled_msg(
-            node, target_model_name, target_model_package, disabled
-        )
-        if disabled:
-            logger.debug(printer.warning_tag(msg))
-        else:
-            dbt.exceptions.warn_or_error(
-                msg,
-                log_fmt=printer.warning_tag('{}')
-            )
-    else:
-        dbt.exceptions.ref_target_not_found(
-            node,
-            target_model_name,
-            target_model_package,
-            disabled=disabled,
-        )
-
-
-def invalid_source_fail_unless_test(
-    node, target_name, target_table_name, disabled
-):
-    if node.resource_type == NodeType.Test:
-        msg = dbt.exceptions.get_source_not_found_or_disabled_msg(
-            node, target_name, target_table_name, disabled
-        )
-        if disabled:
-            logger.debug(printer.warning_tag(msg))
-        else:
-            dbt.exceptions.warn_or_error(
-                msg,
-                log_fmt=printer.warning_tag('{}')
-            )
-    else:
-        dbt.exceptions.source_target_not_found(
-            node,
-            target_name,
-            target_table_name,
-            disabled=disabled
-        )
-
-
-def parse_cli_vars(var_string: str) -> Dict[str, Any]:
-    try:
-        cli_vars = yaml_helper.load_yaml_text(var_string)
-        var_type = type(cli_vars)
-        if var_type is dict:
-            return cli_vars
-        else:
-            type_name = var_type.__name__
-            dbt.exceptions.raise_compiler_error(
-                "The --vars argument must be a YAML dictionary, but was "
-                "of type '{}'".format(type_name))
-    except dbt.exceptions.ValidationException:
-        logger.error(
-            "The YAML provided in the --vars argument is not valid.\n"
-        )
-        raise
 
 
 K_T = TypeVar('K_T')
